requirements.txt
flask==2.3.2
flask_session==0.4.0
numpy==1.24.3
xgboost==2.0.3
llama-cpp-python==0.3.12
joblib==1.3.2
